Move architecture-specific files into their own directories
