+++ .gitignore
@@ -7,10 +7,14 @@
 # Build outputs (TODO build to a subdir so we can exclude that instead)
 /allparserscheckfile
 /basictests
+/benchmark/parse
+/benchmark/perfdiff
+/benchmark/statisticalmodel
 /json2json
 /jsoncheck
 /jsonpointer
 /jsonstats
+/libsimdjson.so*
 /minify
 /numberparsingcheck
 /parse
@@ -25,8 +29,33 @@
 /simdjson.h
 /singleheader/amalgamation_demo
 /singleheader/demo
+/tests/basictests
+/tests/jsoncheck
+/tests/pointercheck
+/tools/json2json
+/tools/jsonstats
+/tools/minify
+# CMake ignore from https://github.com/github/gitignore/blob/master/CMake.gitignore
 
-# Generic from https://github.com/github/gitignore/blob/master/C%2B%2B.gitignore
+CMakeLists.txt.user
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
+Makefile
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
+_deps
+
+# CMake files that may be specific to our installation
+/CPackConfig.cmake
+/CPackSourceConfig.cmake
+# We check in a custom version of root Makefile that is not generated by CMake
+!/Makefile
+
+# C++ ignore from https://github.com/github/gitignore/blob/master/C%2B%2B.gitignore
 
 # Prerequisites
 *.d
