+++ include/CMakeLists.txt
@@ -1,20 +1,20 @@
-set(SIMDJSON_INCLUDE_DIR ${PROJECT_SOURCE_DIR}/include/simdjson)
+set(SIMDJSON_INCLUDE_DIR ${PROJECT_SOURCE_DIR}/include)
 set(SIMDJSON_INCLUDE
-    ${SIMDJSON_INCLUDE_DIR}/common_defs.h
-    ${SIMDJSON_INCLUDE_DIR}/isadetection.h
-    ${SIMDJSON_INCLUDE_DIR}/jsoncharutils.h
-    ${SIMDJSON_INCLUDE_DIR}/jsonformatutils.h
-    ${SIMDJSON_INCLUDE_DIR}/jsonioutil.h
-    ${SIMDJSON_INCLUDE_DIR}/jsonminifier.h
-    ${SIMDJSON_INCLUDE_DIR}/jsonparser.h
-    ${SIMDJSON_INCLUDE_DIR}/numberparsing.h
-    ${SIMDJSON_INCLUDE_DIR}/padded_string.h
-    ${SIMDJSON_INCLUDE_DIR}/parsedjson.h
-    ${SIMDJSON_INCLUDE_DIR}/parsedjsoniterator.h
-    ${SIMDJSON_INCLUDE_DIR}/portability.h
-    ${SIMDJSON_INCLUDE_DIR}/simdjson.h
-    ${SIMDJSON_INCLUDE_DIR}/simdjson_version.h
-    ${SIMDJSON_INCLUDE_DIR}/simdprune_tables.h
-    ${SIMDJSON_INCLUDE_DIR}/stage1_find_marks.h
-    ${SIMDJSON_INCLUDE_DIR}/stage2_build_tape.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/common_defs.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/isadetection.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/jsoncharutils.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonformatutils.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonioutil.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonminifier.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonparser.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/numberparsing.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/padded_string.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/parsedjson.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/parsedjsoniterator.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/portability.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/simdjson.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/simdjson_version.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/simdprune_tables.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/stage1_find_marks.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/stage2_build_tape.h
 )
