+++ include/simdjson/stage1_find_marks.h
@@ -14,6 +14,6 @@ int find_structural_bits(const char *buf, size_t len, simdjson::ParsedJson &pj) 
   return find_structural_bits((const uint8_t *)buf, len, pj);
 }
 
-} // namespace simdjson
+}; // namespace simdjson
 
 #endif
