+++ src/generic/stringparsing.h
