+++ src/stage1_find_marks.cpp
@@ -1,10 +1,10 @@
 #include "simdjson/portability.h"
 
 #ifdef IS_X86_64
-#include "stage1_find_marks_haswell.h"
-#include "stage1_find_marks_westmere.h"
+#include "haswell/stage1_find_marks.h"
+#include "westmere/stage1_find_marks.h"
 #endif // IS_X86_64
 
 #ifdef IS_ARM64
-#include "stage1_find_marks_arm64.h"
+#include "arm64/stage1_find_marks.h"
 #endif // IS_ARM64
