+++ src/stage2_build_tape.cpp
@@ -2,17 +2,17 @@
 
 #ifdef IS_X86_64
 
-#include "stringparsing_haswell.h"
+#include "haswell/stringparsing.h"
 #define TARGETED_ARCHITECTURE Architecture::HASWELL
 #define TARGETED_REGION TARGET_HASWELL
-#include "stage2_build_tape_common.h"
+#include "generic/stage2_build_tape.h"
 #undef TARGETED_ARCHITECTURE
 #undef TARGETED_REGION
 
-#include "stringparsing_westmere.h"
+#include "westmere/stringparsing.h"
 #define TARGETED_ARCHITECTURE Architecture::WESTMERE
 #define TARGETED_REGION TARGET_WESTMERE
-#include "stage2_build_tape_common.h"
+#include "generic/stage2_build_tape.h"
 #undef TARGETED_ARCHITECTURE
 #undef TARGETED_REGION
 
@@ -20,10 +20,10 @@
 
 #ifdef IS_ARM64
 
-#include "stringparsing_arm64.h"
+#include "arm64/stringparsing.h"
 #define TARGETED_ARCHITECTURE Architecture::ARM64
 #define TARGETED_REGION TARGET_ARM64
-#include "stage2_build_tape_common.h"
+#include "generic/stage2_build_tape.h"
 #undef TARGETED_ARCHITECTURE
 #undef TARGETED_REGION
